+++ src/module/nft-exchange/dto/nft-exchange.dto.ts
@@ -119,107 +119,3 @@ export class NftExchangeListRequest {
   @IsString()
   offerorUser?: string;
 }
-
-export class OfferorUserProfile {
-  @ApiProperty()
-  @Expose()
-  walletAddress: string;
-
-  @ApiProperty()
-  @Expose()
-  name: string;
-
-  @ApiProperty()
-  @Expose()
-  image: string;
-
-  static from(entity: UserEntity) {
-    return plainToInstance(this, entity, { excludeExtraneousValues: true });
-  }
-}
-
-export class ExchangeMaterialNFT {
-  @ApiProperty()
-  contractId: number;
-
-  @ApiProperty()
-  name: string;
-
-  @ApiProperty()
-  imageUrl: string;
-
-  @ApiProperty()
-  quantity: number;
-}
-
-export class ExchangeBlueprintOrPuzzleNFT {
-  @ApiProperty()
-  seasonZoneId: number;
-
-  @ApiProperty()
-  seasonName: string;
-
-  @ApiProperty()
-  zoneName: string;
-
-  @ApiProperty()
-  imageUrl: string;
-
-  @ApiProperty()
-  quantity: number;
-}
-
-@ApiExtraModels(
-  OfferorUserProfile,
-  ExchangeMaterialNFT,
-  ExchangeBlueprintOrPuzzleNFT,
-)
-export class NftExchangeListResponse {
-  @ApiProperty()
-  @Expose()
-  nftExchangeOfferId: number;
-
-  @ApiProperty()
-  @Expose()
-  offerorUser: OfferorUserProfile;
-
-  @ApiProperty({
-    oneOf: [
-      { type: 'object', $ref: getSchemaPath(ExchangeMaterialNFT) },
-      { type: 'object', $ref: getSchemaPath(ExchangeBlueprintOrPuzzleNFT) },
-    ],
-  })
-  @Expose()
-  offeredNfts: (ExchangeMaterialNFT | ExchangeBlueprintOrPuzzleNFT)[];
-
-  @ApiProperty({
-    oneOf: [
-      { type: 'object', $ref: getSchemaPath(ExchangeMaterialNFT) },
-      { type: 'object', $ref: getSchemaPath(ExchangeBlueprintOrPuzzleNFT) },
-    ],
-  })
-  @Expose()
-  requestedNfts: (ExchangeMaterialNFT | ExchangeBlueprintOrPuzzleNFT)[];
-
-  @ApiProperty()
-  @Expose()
-  status: NftExchangeOfferStatus;
-
-  @ApiProperty()
-  @Expose()
-  createdAt: Date;
-
-  static from(entity: NftExchangeOfferEntity) {
-    return plainToInstance(
-      this,
-      {
-        ...entity,
-        nftExchangeOfferId: entity.id,
-        offerorUser: OfferorUserProfile.from(entity.offeror),
-      },
-      {
-        excludeExtraneousValues: true,
-      },
-    );
-  }
-}
